Move our functions in register and bootstrap to separate files
