+++ strapi/src/customizations/document-service-middlewares.ts
