+++ strapi/src/index.ts
@@ -1,37 +1,9 @@
 'use strict'
 
 import { Core } from '@strapi/strapi'
-import { registerDocumentServiceMiddlewares } from './document-service-middlewares'
-
-type PermissionSubject = 'api::page.page' | 'api::article.article'
-
-const conditions = [
-  {
-    displayName: 'Document adminGroupId includes starz',
-    name: 'document-admin-group-includes-starz',
-    // category (string, optional): conditions can be grouped into categories available in the admin panel; if undefined, the condition will appear under the "Default" category,
-    plugin: 'content-manager',
-    // The user object passed to handler is not typed unfortunately
-    // See more: https://docs-v4.strapi.io/dev-docs/configurations
-    handler: async (user: any) => {
-      const entitiesWithStarzAdminGroup = await strapi
-        .documents(user.permission.subject as PermissionSubject)
-        .findMany({
-          fields: ['id'],
-          filters: {
-            adminGroups: {
-              adminGroupId: 'starz',
-            },
-          },
-          populate: ['adminGroups'],
-        })
-
-      const entityIds = entitiesWithStarzAdminGroup?.map((entity: any) => entity.id) ?? []
-
-      return { id: { $in: entityIds } }
-    },
-  },
-]
+import { registerDocumentServiceMiddlewares } from './customizations/document-service-middlewares'
+import { bootstrapRevalidateWebhook } from './customizations/bootstrap-revalidate-webhook'
+import { customRbacConditions } from './customizations/custom-rbac-conditions'
 
 export default {
   /**
@@ -55,27 +27,8 @@ export default {
   async bootstrap({ strapi }: { strapi: Core.Strapi }) {
     console.log('Bootstrap function started')
 
-    // create Revalidate webhook according to this suggestion https://github.com/strapi/strapi/pull/20487#issuecomment-2419729414
-    const webhook = await strapi.db.query('strapi::webhook').findOne({
-      where: {
-        name: 'Bootstrapped Revalidate',
-      },
-    })
+    bootstrapRevalidateWebhook({ strapi })
 
-    if (!webhook) {
-      strapi.get('webhookStore').createWebhook({
-        id: 'Bootstrapped Revalidate',
-        name: 'Bootstrapped Revalidate',
-        url: `${process.env.REVALIDATE_NEXT_URL}/api/revalidate?secret=${process.env.REVALIDATE_SECRET_TOKEN}`,
-        events: ['entry.create', 'entry.update', 'entry.publish'],
-        headers: {},
-        enabled: true,
-      })
-      console.log('Revalidate webhook created')
-    } else {
-      console.log('Revalidate webhook already exists')
-    }
-
-    await strapi.admin.services.permission.conditionProvider.registerMany(conditions)
+    await strapi.admin.services.permission.conditionProvider.registerMany(customRbacConditions)
   },
 }
